Reimplement PCB_Process_runBg
On POSIX systems, PCB_Process_runBg uses `vfork` whenever available. If
not, regular `fork` is used, but an additional shared memory region is
attached for the child to signal an error if there is one, since we
can't rely on any exit code for this.
The previously non-existent Windows implementation is now present.